Remove dashes from cluster name
Because I am using that as group names in hosts file and ansible doesn't like
that.

I could name the cluster after its purpose but then I don't know what hosts
are part of that cluster. While that information can be tracked in netbox
I think it may add additional complexity in the way we register hosts/clusters.