Add playbook to register chassis devices in netbox
Chassis devices device_type have subdevice_role set to parent. The devices are
then instantiated from that type like a regular device. After that bays or
slots are created on the device.

The devices that go into the chassis need to be instantiated from device_type
with subdevice_role set to "child" and height set to 0. So some
changes are made in register-system.yaml to reflect that. Additionaly, we need
to specify the rack, parent_device and slot/bay name for the child device.